Unalias git for the PS1
I (probably) have git aliased to hub, but that slows my prompt down
and is super annoying. This seems to be the best solution for now,
though I will look into getting bash completion for the raw hub
command going. Right now, completion of the hub commands are dependent
on being called as a git subcommand.